fixed some bugs and added Execution data to types.go
