+++ litmus-portal/frontend/src/graphql/subscriptions.ts
@@ -11,6 +11,7 @@ export const WORKFLOW_EVENTS_WITH_EXEC_DATA = gql`
       cluster_id
       phase
       execution_data
+      resiliency_score
     }
   }
 `;
