+++ litmus-portal/frontend/src/views/Home/AgentConfiguredHome/RecentWorkflowRuns/WorkflowRunCard/index.tsx
@@ -102,7 +102,10 @@ const WorkflowRunCard: React.FC<WorkflowRunCardProps> = ({ data }) => {
             <Typography
               className={getResiliencyScoreVariant(data.resiliency_score ?? 0)}
             >
-              {data.resiliency_score ? `${data.resiliency_score}%` : 'NA'}
+              {data.resiliency_score === undefined ||
+              data.resiliency_score === null
+                ? 'NA'
+                : `${data.resiliency_score}%`}
             </Typography>
           </div>
