+++ litmus-portal/frontend/src/views/WorkflowDetails/WorkflowInfo/index.tsx
@@ -2,14 +2,15 @@ import { Typography } from '@material-ui/core';
 import { ButtonOutlined } from 'litmus-ui';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
+import TimePopOver from '../../../components/TimePopOver';
 import { ExecutionData } from '../../../models/graphql/workflowData';
 import useStyles from './styles';
-import TimePopOver from '../../../components/TimePopOver';
 
 interface WorkflowInfoProps {
   setIsInfoToggled?: React.Dispatch<React.SetStateAction<boolean>>;
   tab: number;
   data: ExecutionData;
+  resiliency_score?: number;
   cluster_name: string;
 }
 
@@ -17,6 +18,7 @@ const WorkflowInfo: React.FC<WorkflowInfoProps> = ({
   setIsInfoToggled,
   tab,
   data,
+  resiliency_score,
   cluster_name,
 }) => {
   const classes = useStyles();
@@ -56,7 +58,9 @@ const WorkflowInfo: React.FC<WorkflowInfoProps> = ({
           </Typography>
           {/* Static data, will be changed with API response */}
           <Typography className={classes.resiliencyScore}>
-            {!data.resiliency_score ? 'NA' : `${data.resiliency_score}%`}
+            {resiliency_score === undefined || resiliency_score === null
+              ? 'NA'
+              : `${resiliency_score}%`}
           </Typography>
         </div>
